Register resolving the previous commit
It seems I had some problems with uploading some of the changes last time
